Python: Use of init files
This includes a little tidying up to create a Python subsection but the
main area of advice is on avoiding side-effects and heavy initialisation
in init files.

If we start to follow this it should make our code easier to deploy in
different environments and avoid some of the nasty tangles we have in
some of our package initialisation today.